Refactor atc.etl module
