+++ tests/etl/test_extractor.py
@@ -3,7 +3,7 @@ import unittest
 from pyspark.sql import DataFrame
 from pyspark.sql.types import StructType, StructField, IntegerType, StringType
 
-from atc import DelegatingExtractor, Extractor
+from atc.etl import DelegatingExtractor, Extractor
 from atc.spark import Spark
