+++ tests/etl/test_orchestrator_factory.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import MagicMock
 
-from atc import OrchestratorFactory
+from atc.etl import OrchestratorFactory
 
 
 class OrchestratorFactoryTests(unittest.TestCase):
@@ -11,4 +11,8 @@ class OrchestratorFactoryTests(unittest.TestCase):
 
     def test_create_for_multiple_sources_returns_not_none(self):
         sut = OrchestratorFactory.create_for_multiple_sources(MagicMock(), MagicMock(), MagicMock())
+        self.assertIsNotNone(sut)
+
+    def test_create_for_multiple_transformers_returns_not_none(self):
+        sut = OrchestratorFactory.create_for_multiple_transformers(MagicMock(), MagicMock(), MagicMock())
         self.assertIsNotNone(sut)
