+++ tests/etl/test_transformer.py
@@ -3,7 +3,7 @@ import unittest
 from pyspark.sql import DataFrame
 from pyspark.sql.types import StructType, StructField, IntegerType, StringType
 
-from atc import DelegatingTransformer, Transformer
+from atc.etl import DelegatingTransformer, Transformer
 from atc.spark import Spark
